possible softening of the accusatory second person
i suspect this was prompted by some recent event to which i was oblivious; if so, perhaps it would be more civil to drop 2nd person here. or perhaps the intensity here was intentional, in which case leave it by all means. probably there could be better wording here but hopefully it gets the idea across